refactor: move types to shared file
