+++ src/types.ts
@@ -1,0 +1,11 @@
+export type TranslatorInterface = (text: string, priority: number) => Promise<string>;
+
+export interface InnerConfig {
+	isTranslatableNode: (node: Node) => boolean;
+	lazyTranslate: boolean;
+}
+
+export interface Config {
+	isTranslatableNode?: (node: Node) => boolean;
+	lazyTranslate?: boolean;
+}
